# Compiler optimizations