--- a/build_and_run.sh
+++ b/build_and_run.sh
@@ -1,0 +1,15 @@
+#!/bin/bash
+
+# build_and_run.sh
+
+rm -rf build
+rm -rf delphi/cpp/*
+
+mkdir -p build
+cd build
+cmake ..
+cmake --build . -- -j12 DelphiPython
+cp *.so ../delphi/cpp
+cd ..
+
+time python test.py