This pull request includes two simple compiler optimizations:
- a) Use `-O3 -ffast-math` in `CMakeLists.txt`
- b) Inline the `sqr(x) {return x * x}` function from `lib/utils.cpp`

### Testing
I've also added a build and benchmark script, so we can run w/:
```
./build_and_run.sh
```

### Results
| version     | burnin | posterior sampling |
| ----------- | ------ | ------------------ |
| original code | 9.5s | 1.8s |
| a only        | 3.7s | 0.7s |
| a and b       | 1.2s | 0.2s |

So using both optimizations reduces runtime of this example by ~87%.

### Notes
This is a PR to `master`, but the numbers and experiments above are based on comparison to commit `c07eeae9394ab30ca8d984b2ec2e40ab4c2d2e08`, per @manujinda 's recommendation. _I have not tested on `master`_, but it should be easy for you to copy/paste these changes to whichever branch is currently under development.

### Possible Future Work
After these optimizations, profiling via `callgrind` reveals that the majority of runtime is spent computing calls to `exp` in `KDE::pdf`.  To get additional speedups, we'd need to make `exp` run faster.  This may be possible, depending on how much precision is required, by eg. using approximate `exp` functions like the ones described [here](https://stackoverflow.com/questions/10552280).