@bkj 2021-10-01
Testing this some more -- it's possible that `-ffast-math` actually breaks something.  If/before this gets merged -- the impact of `-ffast-math` should be tested.  Speedup w/o `-ffast-math` is still substantial (2.9s).